requirements.txt
openpyxl==3.0.7
python-dotenv==0.19.0
supabase-py==0.0.2
